chore: move discourse api types to shared/types
