+++ src/back/services/update.ts
@@ -2,7 +2,6 @@ import logger from 'decentraland-gatsby/dist/entities/Development/logger'
 import RequestError from 'decentraland-gatsby/dist/entities/Route/error'
 
 import { Discourse } from '../../clients/Discourse'
-import { DiscoursePost } from '../../clients/DiscourseTypes'
 import ProposalModel from '../../entities/Proposal/model'
 import { ProposalAttributes } from '../../entities/Proposal/types'
 import UpdateModel from '../../entities/Updates/model'
@@ -12,6 +11,7 @@ import { inBackground } from '../../helpers'
 import { DiscourseService } from '../../services/DiscourseService'
 import { ErrorService } from '../../services/ErrorService'
 import { FinancialService } from '../../services/FinancialService'
+import { DiscoursePost } from '../../shared/types/discourse'
 import { ErrorCategory } from '../../utils/errorCategories'
 
 import { CoauthorService } from './coauthor'
