+++ src/services/DiscourseService.ts
@@ -4,7 +4,6 @@ import { requiredEnv } from 'decentraland-gatsby/dist/utils/env'
 import { UpdateService } from '../back/services/update'
 import { VoteService } from '../back/services/vote'
 import { Discourse } from '../clients/Discourse'
-import { DiscourseComment, DiscoursePost, DiscoursePostInTopic } from '../clients/DiscourseTypes'
 import ProposalModel from '../entities/Proposal/model'
 import { ProposalWithOutcome } from '../entities/Proposal/outcome'
 import * as proposalTemplates from '../entities/Proposal/templates'
@@ -17,6 +16,7 @@ import { getPublicUpdates, getUpdateUrl } from '../entities/Updates/utils'
 import UserModel from '../entities/User/model'
 import { filterComments } from '../entities/User/utils'
 import { inBackground } from '../helpers'
+import { DiscourseComment, DiscoursePost, DiscoursePostInTopic } from '../shared/types/discourse'
 import { DclProfile } from '../utils/Catalyst/types'
 
 import { ProposalInCreation } from './ProposalService'
