+++ src/services/ProposalService.ts
@@ -5,7 +5,6 @@ import RequestError from 'decentraland-gatsby/dist/entities/Route/error'
 import { DiscordService } from '../back/services/discord'
 import { EventsService } from '../back/services/events'
 import { NotificationService } from '../back/services/notification'
-import { DiscoursePost } from '../clients/DiscourseTypes'
 import { SnapshotProposalContent } from '../clients/SnapshotTypes'
 import UnpublishedBidModel from '../entities/Bid/model'
 import CoauthorModel from '../entities/Coauthor/model'
@@ -18,6 +17,7 @@ import { isGrantProposalSubmitEnabled } from '../entities/Proposal/utils'
 import { SNAPSHOT_SPACE } from '../entities/Snapshot/constants'
 import VotesModel from '../entities/Votes/model'
 import { getEnvironmentChainId } from '../helpers'
+import { DiscoursePost } from '../shared/types/discourse'
 import { getProfile } from '../utils/Catalyst'
 import Time from '../utils/date/Time'
